+++ src/fg_apply_tmpl_16.rs
@@ -107,9 +107,9 @@ unsafe extern "C" fn generate_scaling(
     }
     memset(
         scaling as *mut libc::c_void,
-        (*points.offset(0 as libc::c_int as isize))[1]
+        (*points.offset(0))[1]
             as libc::c_int,
-        (((*points.offset(0 as libc::c_int as isize))[0]
+        (((*points.offset(0))[0]
             as libc::c_int) << shift_x) as libc::c_ulong,
     );
     let mut i: libc::c_int = 0 as libc::c_int;
@@ -197,7 +197,7 @@ pub unsafe extern "C" fn dav1d_prep_grain_16bpc(
         .expect(
             "non-null function pointer",
         )(
-        (*grain_lut.offset(0 as libc::c_int as isize)).as_mut_ptr(),
+        (*grain_lut.offset(0)).as_mut_ptr(),
         data,
         bitdepth_max,
     );
@@ -210,8 +210,8 @@ pub unsafe extern "C" fn dav1d_prep_grain_16bpc(
             .expect(
                 "non-null function pointer",
             )(
-            (*grain_lut.offset(1 as libc::c_int as isize)).as_mut_ptr(),
-            (*grain_lut.offset(0 as libc::c_int as isize)).as_mut_ptr()
+            (*grain_lut.offset(1)).as_mut_ptr(),
+            (*grain_lut.offset(0)).as_mut_ptr()
                 as *const [entry; 82],
             data,
             0 as libc::c_int as intptr_t,
@@ -227,8 +227,8 @@ pub unsafe extern "C" fn dav1d_prep_grain_16bpc(
             .expect(
                 "non-null function pointer",
             )(
-            (*grain_lut.offset(2 as libc::c_int as isize)).as_mut_ptr(),
-            (*grain_lut.offset(0 as libc::c_int as isize)).as_mut_ptr()
+            (*grain_lut.offset(2)).as_mut_ptr(),
+            (*grain_lut.offset(0)).as_mut_ptr()
                 as *const [entry; 82],
             data,
             1 as libc::c_int as intptr_t,
@@ -240,7 +240,7 @@ pub unsafe extern "C" fn dav1d_prep_grain_16bpc(
             (*in_0).p.bpc,
             ((*data).y_points).as_ptr(),
             (*data).num_y_points,
-            (*scaling.offset(0 as libc::c_int as isize)).as_mut_ptr(),
+            (*scaling.offset(0)).as_mut_ptr(),
         );
     }
     if (*data).num_uv_points[0] != 0 {
@@ -248,7 +248,7 @@ pub unsafe extern "C" fn dav1d_prep_grain_16bpc(
             (*in_0).p.bpc,
             ((*data).uv_points[0]).as_ptr(),
             (*data).num_uv_points[0],
-            (*scaling.offset(1 as libc::c_int as isize)).as_mut_ptr(),
+            (*scaling.offset(1)).as_mut_ptr(),
         );
     }
     if (*data).num_uv_points[1] != 0 {
@@ -256,7 +256,7 @@ pub unsafe extern "C" fn dav1d_prep_grain_16bpc(
             (*in_0).p.bpc,
             ((*data).uv_points[1]).as_ptr(),
             (*data).num_uv_points[1],
-            (*scaling.offset(2 as libc::c_int as isize)).as_mut_ptr(),
+            (*scaling.offset(2)).as_mut_ptr(),
         );
     }
     if !((*out).stride[0]
@@ -381,8 +381,8 @@ pub unsafe extern "C" fn dav1d_apply_grain_row_16bpc(
             (*out).stride[0],
             data,
             (*out).p.w as size_t,
-            (*scaling.offset(0 as libc::c_int as isize)).as_ptr(),
-            (*grain_lut.offset(0 as libc::c_int as isize)).as_ptr(),
+            (*scaling.offset(0)).as_ptr(),
+            (*grain_lut.offset(0)).as_ptr(),
             bh,
             row,
             bitdepth_max,
@@ -429,7 +429,7 @@ pub unsafe extern "C" fn dav1d_apply_grain_row_16bpc(
                 (*in_0).stride[1],
                 data,
                 cpw as size_t,
-                (*scaling.offset(0 as libc::c_int as isize)).as_ptr(),
+                (*scaling.offset(0)).as_ptr(),
                 (*grain_lut.offset((1 as libc::c_int + pl) as isize)).as_ptr(),
                 bh_0,
                 row,
